+++ src/JordanMod/patches/ExamplePatch.cs
@@ -4,14 +4,4 @@ namespace JordanMod.Patches;
 
 public class ExamplePatch
 {
-	[HarmonyPatch(typeof(Item), "Start")]
-	[HarmonyPostfix]
-	static void PostItemStartFix(Item __instance)
-	{
-		if (__instance.UIData.itemName.ToLower() == "passport")
-		{
-			__instance.UIData.canDrop = true;
-			__instance.UIData.canThrow = true;
-		}
-	}
 }